Add functionality to the equal button
Add an event listener to the equal button. When it is pressed the working
is split with a regular expression. If the working screen contains two
numbers and an operation seperating the two the correct operation will be
preformed on the two numbers and the answer will be displayed in the answer
screen area.